third_party/WebKit/LayoutTests/imported/wpt/html/browsers/origin/cross-origin-objects/cross-origin-objects-expected.txt

This is a testharness.js-based test.
FAIL Basic sanity-checking assert_equals: Need to run the top-level test from domain {{host}} expected "{{host}}" but got ""
FAIL Only whitelisted properties are accessible cross-origin assert_throws: Should throw when accessing speechSynthesis on Window function "function () { C[prop]; }" did not throw
FAIL [[GetPrototypeOf]] should return null assert_true: cross-origin Window proto is null expected true got false
FAIL [[SetPrototypeOf]] should throw assert_throws: proto set on cross-origin Location function "function () { C.location.__proto__ = new Object(); }" did not throw
PASS [[IsExtensible]] should return true for cross-origin objects 
FAIL [[PreventExtensions]] should throw for cross-origin objects assert_throws: preventExtensions on cross-origin Window should throw function "function () { Object.preventExtensions(C) }" did not throw
PASS [[GetOwnProperty]] - Properties on cross-origin objects should be reported |own| 
FAIL [[GetOwnProperty]] - Property descriptors for cross-origin properties should be set up correctly assert_equals: property descriptor for location should be non-enumerable expected false but got true
FAIL [[Delete]] Should throw on cross-origin objects assert_throws: Can't delete cross-origin indexed property function "function () { delete C[0]; }" did not throw
FAIL [[DefineOwnProperty]] Should throw for cross-origin objects assert_throws: Can't define cross-origin value property length function "function () { Object.defineProperty(obj, prop, valueDesc); }" did not throw
FAIL [[Enumerate]] should return an empty iterator assert_unreached: Shouldn't have been able to enumerate speechSynthesis on cross-origin Window Reached unreachable code
FAIL [[OwnPropertyKeys]] should return all properties from cross-origin objects assert_array_equals: Object.getOwnPropertyNames() gives the right answer for cross-origin Window lengths differ, expected 861 got 13
PASS A and B jointly observe the same identity for cross-origin Window and Location 
PASS Cross-origin functions get local Function.prototype 
FAIL Cross-origin Window accessors get local Function.prototype Cannot read property 'name' of undefined
FAIL Same-origin observers get different functions for cross-origin objects assert_true: same-origin Window functions get their own object expected true got false
FAIL Same-origin observers get different accessors for cross-origin Window assert_true: different Window accessors per-incumbent script settings object expected true got false
FAIL Same-origin observers get different accessors for cross-origin Location assert_true: different Location accessors per-incumbent script settings object expected true got false
Harness: the test ran to completion.
